Allow DefaultMemberPermissions on Commands and make button commands better

+++ GroupBot-Discord/src/main/kotlin/net/dragoncoding/groupbot/discord/command/AbstractDiscordSlashCommand.kt
@@ -6,6 +6,7 @@ import net.dragoncoding.groupbot.common.logging.StatusCode
 import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent
+import net.dv8tion.jda.api.interactions.commands.DefaultMemberPermissions
 import net.dv8tion.jda.api.interactions.commands.build.OptionData
 
 abstract class AbstractDiscordSlashCommand : IDiscordSlashCommand {
@@ -19,7 +20,10 @@ abstract class AbstractDiscordSlashCommand : IDiscordSlashCommand {
     override val acknowledgeButton: Boolean = false
     override val acknowledgeModal: Boolean = false
 
+    override var deleteOriginal: Boolean = false
+
     override fun getCommandOptions(): List<OptionData> = listOf()
+    override fun getCommandPermission(): DefaultMemberPermissions? = null
 
     override fun runButton(event: ButtonInteractionEvent): IStatus =
         Status(StatusCode.ERROR, "No Button Event implemented")
@@ -29,16 +33,16 @@ abstract class AbstractDiscordSlashCommand : IDiscordSlashCommand {
 
     override fun onCommand(event: SlashCommandInteractionEvent): IStatus {
         if (acknowledgeCommand)
-            event.deferReply(true).queue()
+            event.deferReply(true).complete()
 
         val status = runCommand(event)
 
         if (status.isOk) {
-            if (acknowledgeCommand) {
+            if (acknowledgeCommand && deleteOriginal) {
                 event.hook.deleteOriginal().queue()
             }
         } else {
-            val message = "An error occurred while executing this command: ${status.message}"
+            val message = "An error occurred while executing this command:\n${status.message}"
             if (acknowledgeCommand)
                 event.hook.editOriginal(message).queue()
             else
@@ -49,16 +53,16 @@ abstract class AbstractDiscordSlashCommand : IDiscordSlashCommand {
 
     override fun onButton(event: ButtonInteractionEvent): IStatus {
         if (acknowledgeButton)
-            event.deferReply(true).queue()
+            event.deferReply(true).complete()
 
         val status = runButton(event)
 
         if (status.isOk) {
-            if (acknowledgeButton) {
+            if (acknowledgeButton && deleteOriginal) {
                 event.hook.deleteOriginal().queue()
             }
         } else {
-            val message = "An error occurred while processing this button: ${status.message}"
+            val message = "An error occurred while processing this button:\n${status.message}"
             if (acknowledgeButton)
                 event.hook.editOriginal(message).queue()
             else
@@ -69,16 +73,16 @@ abstract class AbstractDiscordSlashCommand : IDiscordSlashCommand {
 
     override fun onModal(event: ModalInteractionEvent): IStatus {
         if (acknowledgeModal)
-            event.deferReply(true).queue()
+            event.deferReply(true).complete()
 
         val status = runModal(event)
 
         if (status.isOk) {
-            if (acknowledgeModal) {
+            if (acknowledgeModal && deleteOriginal) {
                 event.hook.deleteOriginal().queue()
             }
         } else {
-            val message = "An error occurred while processing this modal: ${status.message}"
+            val message = "An error occurred while processing this modal:\n${status.message}"
             if (acknowledgeModal)
                 event.hook.editOriginal(message).queue()
             else
